refactor: migrate basic group
