+++ packages/webpack-cli/lib/webpack-cli.js
@@ -1,4 +1,6 @@
+const path = require('path');
 const { options } = require('colorette');
+const webpackMerge = require('webpack-merge');
 const GroupHelper = require('./utils/GroupHelper');
 const handleConfigResolution = require('./groups/ConfigGroup');
 const resolveMode = require('./groups/resolveMode');
@@ -6,7 +8,7 @@ const resolveStats = require('./groups/resolveStats');
 const resolveOutput = require('./groups/resolveOutput');
 const { Compiler } = require('./utils/Compiler');
 const { groups, core } = require('./utils/cli-flags');
-const webpackMerge = require('webpack-merge');
+const basicResolver = require('./groups/basicResolver');
 const { toKebabCase } = require('./utils/helpers');
 const argParser = require('./utils/arg-parser');
 const { outputStrategy } = require('./utils/merge-strategies');
@@ -106,11 +108,6 @@ class WebpackCLI extends GroupHelper {
     resolveGroups() {
         for (const [key, value] of this.groupMap.entries()) {
             switch (key) {
-                case groups.BASIC_GROUP: {
-                    const BasicGroup = require('./groups/BasicGroup');
-                    this.basicGroup = new BasicGroup(value);
-                    break;
-                }
                 case groups.ADVANCED_GROUP: {
                     const AdvancedGroup = require('./groups/AdvancedGroup');
                     this.advancedGroup = new AdvancedGroup(value);
@@ -217,12 +214,7 @@ class WebpackCLI extends GroupHelper {
      * @returns {void}
      */
     _handleDefaultEntry() {
-        if (!this.basicGroup) {
-            const BasicGroup = require('./groups/BasicGroup');
-            this.basicGroup = new BasicGroup();
-        }
-        const defaultEntry = this.basicGroup.resolveFilePath(null, 'index.js');
-        const options = { entry: defaultEntry };
+        const options = { entry: path.resolve('index.js') };
         this._mergeOptionsToConfiguration(options);
     }
 
@@ -240,7 +232,7 @@ class WebpackCLI extends GroupHelper {
             .then(() => this._baseResolver(resolveMode, parsedArgs, this.compilerConfiguration))
             .then(() => this._baseResolver(resolveOutput, parsedArgs, {}, outputStrategy))
             .then(() => this._handleCoreFlags())
-            .then(() => this._handleGroupHelper(this.basicGroup))
+            .then(() => this._baseResolver(basicResolver, parsedArgs))
             .then(() => this._handleGroupHelper(this.advancedGroup))
             .then(() => this._baseResolver(resolveStats, parsedArgs))
             .then(() => this._handleGroupHelper(this.helpGroup));
